refactor: packages/contracts・packages/schema を廃止しスキーマを各アプリに移動
一人開発では中間パッケージの維持コストが価値を上回るため整理する。

- packages/contracts 削除: AppType の単純な re-export 層を廃止
  - mobile/lib/api.ts を @animeishi/api 直接参照に変更
- packages/schema 削除: スキーマを使う側に移動
  - signInSchema / signUpSchema → apps/mobile/lib/validators.ts
  - profileUpdateSchema → services/api/src/schema/validators.ts
  - errors.ts はデッドコードのため破棄
- テスト追加: 移動したスキーマの単体テストを各アプリに作成
  - apps/mobile/__tests__/validators.test.ts
  - services/api/src/schema/__tests__/validators.test.ts
- vitest.eslint.config.ts → vitest.node.config.ts にリネーム
  スキーマテスト（Workers非依存）を同configで実行するよう拡張
