+++ apps/mobile/jest.config.js
@@ -7,9 +7,7 @@ module.exports = {
   ],
   moduleNameMapper: {
     "^@/(.*)$": "<rootDir>/$1",
-    "^@animeishi/schema$": "<rootDir>/../../packages/schema/src/index.ts",
-    "^@animeishi/contracts$":
-      "<rootDir>/../../packages/contracts/src/index.ts",
+    "^@animeishi/api$": "<rootDir>/../../services/api/src/index.ts",
     "^(\\.{1,2}/.*)\\.js$": "$1",
     "^@clerk/clerk-expo$": "<rootDir>/__mocks__/@clerk/clerk-expo.ts",
   },
