+++ apps/mobile/lib/api.ts
@@ -1,5 +1,5 @@
 import { hc } from "hono/client";
-import type { AppType } from "@animeishi/contracts";
+import type { AppType } from "@animeishi/api";
 
 const apiUrl = process.env.EXPO_PUBLIC_API_URL ?? "http://localhost:8787";
