+++ apps/mobile/package.json
@@ -15,8 +15,7 @@
     "format:check": "prettier --check app lib store"
   },
   "dependencies": {
-    "@animeishi/contracts": "workspace:*",
-    "@animeishi/schema": "workspace:*",
+    "@animeishi/api": "workspace:*",
     "@clerk/clerk-expo": "^2.10.0",
     "@expo/metro-runtime": "^6.1.2",
     "@expo/vector-icons": "^15.1.1",
@@ -44,6 +43,7 @@
     "react-native-screens": "~4.16.0",
     "react-native-worklets": "0.8.1",
     "tailwindcss": "^3.4.17",
+    "zod": "^4.3.6",
     "zustand": "^5.0.4"
   },
   "devDependencies": {
