--- packages/contracts/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@animeishi/contracts",
-  "version": "0.0.1",
-  "private": true,
-  "types": "./src/index.ts",
-  "scripts": {
-    "build": "tsc -p tsconfig.json",
-    "lint": "eslint src",
-    "format": "prettier --write src",
-    "format:check": "prettier --check src"
-  },
-  "dependencies": {
-    "@animeishi/api": "workspace:*",
-    "hono": "^4.7.10"
-  },
-  "devDependencies": {
-    "@cloudflare/workers-types": "^4.20260416.2",
-    "@types/node": "^25.6.0",
-    "eslint": "^10.2.0",
-    "prettier": "^3.5.3",
-    "typescript": "^6.0.2",
-    "typescript-eslint": "^8.58.2"
-  }
-}
