--- packages/contracts/src/index.ts
@@ -1,1 +1,0 @@
-export type { AppType } from "@animeishi/api";
