--- packages/schema/acceptance/01_auth.md
@@ -1,39 +1,0 @@
-# 認証 受け入れ基準
-
-## サインアップ
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| A-01 | 有効なメール・パスワード・ユーザー名でサインアップ | アカウント作成成功、`profiles` に行が挿入される |
-| A-02 | 既存メールアドレスでサインアップ | `email-already-in-use` エラー表示 |
-| A-03 | 無効なメールアドレス形式 | バリデーションエラー「有効なメールアドレスを入力してください」 |
-| A-04 | 7文字以下のパスワード | バリデーションエラー「パスワードは8文字以上で入力してください」 |
-| A-05 | 英字のみのパスワード | バリデーションエラー「パスワードは英字と数字を含む必要があります」 |
-| A-06 | 脆弱パターン（password1）のパスワード | バリデーションエラー「より安全なパスワードを設定してください」 |
-| A-07 | パスワードと確認が不一致 | バリデーションエラー「パスワードが一致しません」 |
-| A-08 | 1文字のユーザー名 | バリデーションエラー「ユーザー名は2文字以上で入力してください」 |
-| A-09 | 予約語 (admin) のユーザー名 | バリデーションエラー「別のユーザー名を選択してください」 |
-
-## サインイン
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| A-10 | 登録済みメール・正しいパスワード | サインイン成功、ホーム画面へ遷移 |
-| A-11 | 未登録メールアドレス | `user-not-found` エラー表示 |
-| A-12 | 間違ったパスワード | `wrong-password` エラー表示 |
-| A-13 | 空のメールアドレス | バリデーションエラー「メールアドレスを入力してください」 |
-| A-14 | 空のパスワード | バリデーションエラー「パスワードを入力してください」 |
-
-## パスワードリセット
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| A-15 | 登録済みメールアドレスでリセット申請 | リセットメール送信成功メッセージ表示 |
-| A-16 | 未登録メールアドレスでリセット申請 | セキュリティ上、成功と同じメッセージを表示（ユーザー列挙防止） |
-
-## アカウント削除
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| A-17 | 認証済みユーザーがアカウント削除 | `profiles` 行が削除、関連データ CASCADE 削除、サインアウト |
-| A-18 | 未認証状態でアカウント削除 API 呼び出し | 401 Unauthorized |
