--- packages/schema/acceptance/02_anime.md
@@ -1,41 +1,0 @@
-# アニメ・視聴履歴・お気に入り 受け入れ基準
-
-## アニメ一覧 / 検索 / ソート
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| AN-01 | アニメ一覧画面を開く | `GET /titles` を呼び出し、タイトル一覧が表示される |
-| AN-02 | オフライン時に一覧画面を開く | TanStack Query / AsyncStorage のキャッシュから表示 |
-| AN-03 | タイトル名で検索 | 前方一致・部分一致でフィルタされる |
-| AN-04 | 読み仮名 (yomi) で検索 | ひらがな入力で対応タイトルがヒットする |
-| AN-05 | 空文字で検索 | 全件表示 |
-| AN-06 | タイトル昇順でソート | 50音順に並ぶ |
-| AN-07 | 放映年でソート（新しい順） | firstYear 降順に並ぶ |
-| AN-08 | ジャンルでフィルタ | 選択ジャンルのみ表示 |
-| AN-09 | 複数ジャンルで AND フィルタ | すべてのジャンルに属するタイトルのみ |
-| AN-10 | 検索中にキーボードを閉じる | 結果を維持したまま画面が動かない |
-| AN-11 | 検索結果 0 件 | 「該当するアニメが見つかりませんでした」を表示 |
-| AN-12 | `GET /titles` が失敗 | `fetch_failed` エラーメッセージを表示、再試行ボタンを表示 |
-
-## 視聴履歴 / Watch 画面
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| W-01 | 未視聴アニメをウォッチリストに追加 | `POST /me/watch-histories` 成功、リストに表示 |
-| W-02 | 視聴済みアニメを再追加しようとする | 追加 UI が非活性または「既に追加済み」と表示 |
-| W-03 | 視聴リストからアニメを削除 | `DELETE /me/watch-histories/:tid` 成功、リストから消える |
-| W-04 | 未認証でウォッチリスト取得 | 401 Unauthorized |
-| W-05a | 未認証で他ユーザーの watch-histories を取得 | 401 Unauthorized |
-| W-05b | 認証済みだが他ユーザーの watch-histories を取得 | 403 Forbidden |
-
-## お気に入り
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| F-01 | アニメをお気に入りに追加 | `POST /me/favorites` 成功、星マークが活性化 |
-| F-02 | お気に入り済みアニメの星マークをタップ | `DELETE /me/favorites/:tid` 成功、星マークが非活性 |
-| F-03 | お気に入り一覧を開く | `GET /me/favorites` が呼ばれ、一覧表示 |
-| F-04 | 未認証でお気に入り取得 | 401 Unauthorized |
-| F-05a | 未認証で他ユーザーのお気に入りを取得 | 401 Unauthorized |
-| F-05b | 認証済みだが他ユーザーのお気に入りを取得 | 403 Forbidden |
-| F-06 | お気に入りが 0 件 | 「まだお気に入りはありません」を表示 |
