--- packages/schema/acceptance/03_social.md
@@ -1,86 +1,0 @@
-# ソーシャル・QR・プロフィール・公開ページ・Gemini 受け入れ基準
-
-## QR コード生成
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| QR-01 | 認証済みユーザーが名刺画面を開く | 自分の userId を埋め込んだ QR コードが表示される |
-| QR-02 | 生成された QR の内容 | `animeishi://user/{userId}` フォーマット |
-| QR-03 | QR を旧フォーマット (https://viewer) でも生成できる | 設定に応じて URL フォーマット切替が可能 |
-| QR-04 | userId が 28文字の Firebase UID の場合 | QR データが正しく生成される |
-| QR-05 | userId が Clerk ID (user_xxx) の場合 | QR データが正しく生成される |
-| QR-06 | QR コード画像のシェア | ネイティブシェアシートが起動する |
-
-## QR スキャン / フレンド追加
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| SC-01 | `animeishi://user/{uid}` QR をスキャン | userId が正しく解析される |
-| SC-02 | `https://...viewer.../user/{uid}` QR をスキャン | userId が正しく解析される |
-| SC-03 | 28文字生 UID の QR をスキャン | userId が正しく解析される |
-| SC-04 | 自分の QR をスキャン | `self_scan` エラー表示 |
-| SC-05 | 既にフレンドの QR をスキャン | `already_friend` エラー表示 |
-| SC-06 | 存在しないユーザーの QR をスキャン | `user_not_found` エラー表示 |
-| SC-07 | 無効な QR（非対応フォーマット）をスキャン | `invalid_format` エラー表示 |
-| SC-08 | カメラ権限なし | `camera_permission` エラー + 設定画面へのリンク表示 |
-| SC-09 | フレンド追加成功 | `POST /me/friends` 成功、双方の `friendships` に行が挿入される |
-| SC-10 | カメラタブ非フォーカス時 | `expo-camera` が unmount されリソースを解放する |
-
-## SNS / フレンド一覧 / 削除
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| SN-01 | SNS 画面を開く | フレンドの視聴履歴タイムラインが表示される |
-| SN-02 | フレンド一覧を開く | `GET /me/friends` が呼ばれ、N+1 なく一覧表示 |
-| SN-03 | フレンドを削除 | `DELETE /me/friends/:id` 成功、片方向削除 |
-| SN-04 | フレンドが 0 人 | 「まだフレンドがいません。QRコードをスキャンして追加しましょう」表示 |
-| SN-05 | 未認証でフレンド一覧取得 | 401 Unauthorized |
-| SN-06a | 未認証で他ユーザーのフレンドを取得 | 401 Unauthorized |
-| SN-06b | 認証済みだが他ユーザーのフレンドを取得 | 403 Forbidden |
-
-## プロフィール編集 / 画像
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| PR-01 | プロフィール画面を開く | 現在のユーザー情報が表示される |
-| PR-02 | ユーザー名を有効な値に変更 | `PUT /me/profile` 成功、表示が更新される |
-| PR-03 | 21文字以上のユーザー名 | バリデーションエラー「ユーザー名は20文字以内で入力してください」 |
-| PR-04 | プロフィール画像をアップロード | `expo-image-manipulator` で WebP/512px/q70 圧縮後 R2 に PUT、50KB 以下 |
-| PR-05 | プロフィール画像をカメラで撮影 | 撮影後に同様の圧縮処理が走る |
-| PR-06 | 公開設定を非公開に変更 | `is_public = false` に更新、公開ページからアクセス不可になる |
-| PR-07 | 選択ジャンルを変更 | `selectedGenres` が更新される |
-| PR-08 | 無効なジャンルを含む更新 | バリデーションエラー「有効なジャンルを選択してください」 |
-| PR-09 | 500文字超のコメント | バリデーションエラー「コメントは500文字以内で入力してください」 |
-| PR-10 | 未認証でプロフィール更新 | 401 Unauthorized |
-
-## 名刺 (Home)
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| ME-01 | ホーム画面を開く | 自分の名刺（プロフィール + 画像 + QR）が表示される |
-| ME-02 | 名刺画像をアップロード | R2 に保存、`profiles.meishi_image_key` が更新される |
-| ME-03 | 名刺をピンチズームで拡大 | `react-native-gesture-handler` でジェスチャー反応 |
-| ME-04 | ビューア URL を共有 | `https://{host}/user/{userId}` がシェアシートに渡される |
-| ME-05 | 名刺がない状態 | 「名刺画像を設定してください」プレースホルダー表示 |
-| ME-06 | Gemini 傾向分析ボタンタップ | `/analysis/gemini` を呼び出し、結果を名刺に表示 |
-
-## Viewer 公開ページ
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| VI-01 | 未ログインで `/user/{uid}` にアクセス | `is_public = true` ならプロフィール + 視聴履歴が表示される |
-| VI-02 | `is_public = false` のユーザーにアクセス | 「非公開ページです」と表示 |
-| VI-03 | OGP タグの確認 | `og:title`, `og:image` が HTML に含まれる（Hono SSR） |
-| VI-04 | 存在しない UID でアクセス | 404 ページ表示 |
-| VI-05 | クローラが `Accept: text/html` でアクセス | Hono が OGP 付き静的 HTML を返す |
-
-## Gemini 傾向分析
-
-| ID | 入力 | 期待結果 |
-|---|---|---|
-| GM-01 | 視聴済みアニメが 1 件以上の状態で分析実行 | `POST /analysis/gemini` 成功、出力は 80〜120 文字のコメントが返る |
-| GM-02 | 視聴済みアニメが 0 件の状態で分析実行 | バリデーションエラーまたは「アニメを視聴してから実行してください」表示 |
-| GM-03 | Gemini API タイムアウト | フォールバックメッセージを表示 |
-| GM-04 | 未認証で分析 API 呼び出し | 401 Unauthorized |
-| GM-05 | レート制限（同一ユーザーが短時間に連打） | 429 Too Many Requests |
-| GM-06 | 分析結果に不適切な内容が含まれる | クライアント側フィルタまたはフォールバックメッセージを表示 |
