--- packages/schema/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@animeishi/schema",
-  "version": "0.0.1",
-  "private": true,
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "build": "tsc -p tsconfig.json",
-    "test": "vitest run",
-    "test:watch": "vitest",
-    "lint": "eslint src __tests__",
-    "format": "prettier --write src __tests__",
-    "format:check": "prettier --check src __tests__"
-  },
-  "dependencies": {
-    "zod": "^4.3.6"
-  },
-  "devDependencies": {
-    "@types/node": "^25.6.0",
-    "eslint": "^10.2.0",
-    "prettier": "^3.5.3",
-    "typescript": "^6.0.2",
-    "typescript-eslint": "^8.58.2",
-    "vitest": "^4.1.4"
-  }
-}
