--- packages/schema/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./validators";
-export * from "./errors";
