--- packages/schema/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2022",
-    "module": "ESNext",
-    "moduleResolution": "Bundler",
-    "strict": true,
-    "skipLibCheck": true,
-    "declaration": true,
-    "outDir": "./dist"
-  },
-  "include": ["src", "__tests__"],
-  "exclude": ["node_modules", "dist"]
-}
