+++ services/api/package.json
@@ -8,9 +8,9 @@
     "dev": "wrangler dev",
     "build": "tsc --noEmit",
     "deploy": "wrangler deploy",
-    "test": "vitest run && vitest run --config vitest.eslint.config.ts",
+    "test": "vitest run && vitest run --config vitest.node.config.ts",
     "test:watch": "vitest",
-    "test:eslint-rules": "vitest run --config vitest.eslint.config.ts",
+    "test:node": "vitest run --config vitest.node.config.ts",
     "lint": "eslint src __tests__",
     "format": "prettier --write src __tests__",
     "format:check": "prettier --check src __tests__",
@@ -20,7 +20,6 @@
     "db:studio": "drizzle-kit studio"
   },
   "dependencies": {
-    "@animeishi/schema": "workspace:*",
     "@clerk/backend": "^3.2.12",
     "@clerk/hono": "^0.1.15",
     "@hono/zod-validator": "^0.7.6",
