+++ services/api/vitest.config.ts
@@ -20,6 +20,10 @@ export default defineConfig({
     globals: true,
     // ESLint の RuleTester は Node.js ネイティブモジュールを使うため
     // Workers プールから除外し、Node.js プール（デフォルト）で実行する
-    exclude: ["**/__tests__/no-direct-db.test.js", "node_modules/**"],
+    exclude: [
+      "**/__tests__/no-direct-db.test.js",
+      "src/schema/__tests__/**",
+      "node_modules/**",
+    ],
   },
 });
