--- services/api/vitest.eslint.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from "vitest/config";
-
-// ESLint カスタムルールのテスト用設定（Node.js プールで実行）
-export default defineConfig({
-  test: {
-    globals: true,
-    environment: "node",
-    include: ["**/__tests__/no-direct-db.test.js"],
-  },
-});
